# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Call:
    properties:
      customerEmotion:
        type: string
      customerId:
        type: integer
      duration:
        type: string
      employeeEmotion:
        type: string
      employeeId:
        type: integer
      id:
        type: integer
      overallEmotion:
        type: string
      segments:
        items:
          $ref: '#/definitions/model.Segment'
        type: array
      startTime:
        type: string
    type: object
  model.CallTimeFilterPayload:
    properties:
      endTime:
        type: string
      startTime:
        type: string
      userId:
        type: integer
    type: object
...
  model.Customer:
    properties:
      calls:
        items:
          $ref: '#/definitions/model.Call'
        type: array
      id:
        type: integer
      phone:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: integer
    type: object
  model.Employee:
    properties:
      calls:
        items:
          $ref: '#/definitions/model.Call'
        type: array
      id:
        type: integer
      name:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: integer
    type: object
...
  model.Response:
    properties:
      data:
        type: object
      message:
        type: string
      success:
        type: boolean
    type: object
  model.Segment:
    properties:
      callId:
        type: integer
      emotion:
        type: string
      id:
        type: integer
      segNum:
        type: integer
    type: object
  model.User:
    properties:
      companyName:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
host: localhost:19001
info:
  contact: {}
  description: SPSER - Speech Emotion Recognition Service.
  title: Speech Emotion Recognition APIs
  version: "1.0"
paths:
...
  /customer/{id}:
    get:
      consumes:
      - application/json
      description: Get one customer
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get one customer
      tags:
      - customer-manager-apis
...
  /customer/calls:
    get:
      consumes:
      - application/json
      description: Get all calls of an user with id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all calls of an user with id
      tags:
      - customer-manager-apis
  /customer/calls/filter:
    put:
      consumes:
      - application/json
      description: model.CallTimeFilterPayload => []Call
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CallTimeFilterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all calls in timeframe of an user with id
      tags:
      - customer-manager-apis
  /customer/create:
    post:
      consumes:
      - application/json
      description: Get all customers
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all customers
      tags:
      - customer-manager-apis
  /customer/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete one customer
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete one customer
      tags:
      - customer-manager-apis
  /customer/update:
    put:
      consumes:
      - application/json
      description: Update one customer
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update one customer
      tags:
      - customer-manager-apis
  /employee/{id}:
    get:
      consumes:
      - application/json
      description: Get one employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get one employee
      tags:
      - employee-manager-apis
...
  /employee/calls:
    get:
      consumes:
      - application/json
      description: Get all calls of an user with id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all calls of an user with id
      tags:
      - employee-manager-apis
  /employee/create:
    post:
      consumes:
      - application/json
      description: Get all employees
      parameters:
      - description: employee
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/model.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all employees
      tags:
      - employee-manager-apis
  /employee/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete one employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete one employee
      tags:
      - employee-manager-apis
  /employee/filter/calls:
    put:
      consumes:
      - application/json
      description: model.CallTimeFilterPayload => []Call
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CallTimeFilterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all calls in timeframe of an user with id
      tags:
      - employee-manager-apis
  /employee/update:
    put:
      consumes:
      - application/json
      description: Update one employee
      parameters:
      - description: employee
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/model.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update one employee
      tags:
      - employee-manager-apis
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"